+++ compile/index.js
@@ -1,22 +1,22 @@
-const { _myNewPackage } = require('./mnp')
+const { _yarnS } = require('./yarn-s')
 
 /**
- * {{ description }}
- * @param {!_myNewPackage.Config} config Options for the program.
+ * Run Multiple Yarn Commands In Series.
+ * @param {!_yarnS.Config} config Options for the program.
  * @param {boolean} [config.shouldRun=true] A boolean option. Default `true`.
  * @param {string} [config.text] A text to return.
  * @return {Promise<string>}
  */
 function myNewPackge(config) {
-  return _myNewPackage(config)
+  return _yarnS(config)
 }
 
 module.exports = myNewPackge
 
 /* typal types/index.xml namespace */
 /**
- * @typedef {_myNewPackage.Config} Config `＠record` Options for the program.
- * @typedef {Object} _myNewPackage.Config `＠record` Options for the program.
+ * @typedef {_yarnS.Config} Config `＠record` Options for the program.
+ * @typedef {Object} _yarnS.Config `＠record` Options for the program.
  * @prop {boolean} [shouldRun=true] A boolean option. Default `true`.
  * @prop {string} [text] A text to return.
  */