+++ src/index.js
@@ -1,21 +1,21 @@
 import { c } from 'erte'
 
 /**
- * {{ description }}
- * @param {!_myNewPackage.Config} [config] Options for the program.
+ * Run Multiple Yarn Commands In Series.
+ * @param {!_yarnS.Config} [config] Options for the program.
  * @param {boolean} [config.shouldRun=true] A boolean option. Default `true`.
  * @param {string} [config.text] A text to return.
  */
-export default async function myNewPackage(config = {}) {
+export default async function yarnS(config = {}) {
   const {
     shouldRun = true,
     text = '',
   } = config
   if (!shouldRun) return
-  console.log('my-new-package called with %s', c(text, 'yellow'))
+  console.log('yarn-s called with %s', c(text, 'yellow'))
   return text
 }
 
 /**
- * @typedef {import('../types').Config} _myNewPackage.Config
+ * @typedef {import('../types').Config} _yarnS.Config
  */